requirements.txt
requests
beautifulsoup4
lxml
jq
python-dotenv